server/requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
configparser==3.5.0
Django==1.11.6
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.7.0
enum34==1.1.6
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
MySQL-python==1.2.5
mysqlclient==1.3.12
pylint==1.7.4
pytz==2017.2
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
